Code refactored and restructred according atomic design
