+++ src/App.js
@@ -2,11 +2,11 @@ import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 import Layout from "./components/Layout";
 import PrivateRoute from "./components/PrivateRoute";
 import { UserProvider } from "./contexts/Context";
-import Answers from "./pages/Answers";
-import Home from "./pages/Home";
+import AnswersPage from "./pages/AnswersPage";
+import HomePage from "./pages/HomePage";
 import NotFound from "./pages/NotFound";
-import { Questions } from "./pages/Questions";
-import Quiz from "./pages/Quiz";
+import { QuestionsPage } from "./pages/QuestionsPage";
+import QuizPage from "./pages/QuizPage";
 import Signin from "./pages/Signin";
 
 function App() {
@@ -16,10 +16,10 @@ function App() {
         <Layout>
           <Switch>
             <Route exact path={"/signin"} component={Signin} />
-            <PrivateRoute exact path={"/"} component={Home} />
-            <PrivateRoute exact path={"/questions"} component={Questions} />
-            <PrivateRoute exact path={"/answers"} component={Answers} />
-            <PrivateRoute exact path={"/quiz"} component={Quiz} />
+            <PrivateRoute exact path={"/"} component={HomePage} />
+            <PrivateRoute exact path={"/questions"} component={QuestionsPage} />
+            <PrivateRoute exact path={"/answers"} component={AnswersPage} />
+            <PrivateRoute exact path={"/quiz"} component={QuizPage} />
             <Route component={NotFound} />
           </Switch>
         </Layout>
