+++ src/pages/AnswersPage/style/index.css
@@ -1,0 +1,3 @@
+.question-collapse {
+  background: #caebe2;
+}
