+++ src/pages/QuizPage/hooks/useQuiz.js
@@ -1,11 +1,16 @@
 import { useEffect, useState } from "react";
-import { createAnswer, getAnswersByUserId } from "../mock-data/api";
-import { storageData } from "../utils/storageData";
+import {
+  createAnswer,
+  getAllQuestions,
+  getAnswersByUserId,
+} from "../../../mock-data/api";
+import { storageData } from "../../../utils/storageData";
 
-export default function useAnswers(userId) {
+export default function useQuiz(userId) {
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(false);
   const [answers, setAnswers] = useState(null);
+  const [questions, setQuestions] = useState([]);
 
   async function fetchAnswers() {
     setLoading(true);
@@ -19,8 +24,22 @@ export default function useAnswers(userId) {
         setError(error.toString());
       });
   }
+  async function fetchAllQuestions() {
+    setLoading(true);
+    await getAllQuestions()
+      .then((res) => {
+        storageData.setValue("questions", res);
+        setQuestions(res);
+        setLoading(false);
+      })
+      .catch((error) => {
+        setLoading(false);
+        setError(error.toString());
+      });
+  }
   useEffect(() => {
     fetchAnswers();
+    fetchAllQuestions();
   }, [userId]);
 
   async function submitAnswer(userId, data) {
@@ -43,5 +62,6 @@ export default function useAnswers(userId) {
     error,
     answers,
     submitAnswer,
+    questions,
   };
 }
