+++ src/pages/QuizPage/style/index.css
