+++ src/pages/Signin/style/index.css
@@ -1,4 +1,5 @@
 .signin-form {
+  background: rgba(255, 255, 255, 0.7);
   width: 400px;
   position: absolute;
   top: 40%;
